minor renaming/refactoring
