+++ constants.py
@@ -19,3 +19,6 @@ FLAIR_ID = "a3994b2e-d384-11ea-bf32-0e7e74729027"  # News flair
 
 SNO_ADMIN_USERNAME = "di_admin"
 SNO_ADMIN_PASSWORD = os.environ.get("SNO_ADMIN_PASSWORD", None)
+
+COPY_EDIT_GCAL_ID = "c_7f9830c5fef0310931ee81c0c61b63bb05612190984b8bc15652a34bffffa618@group.calendar.google.com" # Google Calendar ID for copy editing calendar
+COPY_CHIEF_EMAIL = "***REMOVED***" # For copy editing GCal
